% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_modmed.R
\name{complex_modmed}
\alias{complex_modmed}
\title{Title: Moderated Mediation test of multilevel model (random-random; moderator at level two)
This function can calculate moderated mediation effect of multi-level model that
 contains two random paths.}
\usage{
complex_modmed(
  Model,
  first_stage_mod = TRUE,
  first_path = NULL,
  second_path = NULL,
  w = NULL,
  output = "Untitled",
  rep = 20000,
  conf = 95
)
}
\arguments{
\item{Model}{list object containing the parameters of the model
(from Mplus output file)}

\item{first_stage_mod}{a binary variable indicating whether the moderation
effect is on the first path (default: TRUE)}

\item{first_path}{the name of the first random path (should not be 'fixed')}

\item{second_path}{the name of the second random path (should not be 'fixed')}

\item{w}{the name of the moderator}

\item{output}{the name of the output image file (.png)}

\item{rep}{the number of repetitions (default: 20000)}

\item{conf}{the critical value used to generate confidence interval}
}
\value{
An image containing the distribution of the indirect effect and
confidence interval
}
\description{
Title: Moderated Mediation test of multilevel model (random-random; moderator at level two)
This function can calculate moderated mediation effect of multi-level model that
 contains two random paths.
}
