% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_version.R
\name{complex_version}
\alias{complex_version}
\title{Title: Complex version of mediation test}
\usage{
complex_version(
  Model,
  first_path = NULL,
  second_path = NULL,
  output = "Untitled",
  rep = 20000,
  conf = 95
)
}
\arguments{
\item{Model}{a list object containing the parameters of the model
(from Mplus output file)}

\item{first_path}{the name of the first random path (should not be 'fixed')}

\item{second_path}{the name of the second random path (should not be 'fixed')}

\item{output}{the name of the output image file (.png)}

\item{rep}{the number of repetitions (default: 20000)}

\item{conf}{the critical value used to generate confidence interval}
}
\value{
An image containing the distribution of the indirect effect and
confidence interval
}
\description{
The template of the monte carlo method in this function is from
 https://quantpsy.org/medmc/medmc.htm.
This function can handle multi-level model
 (1-1-1 mediation with two random paths).
}
\examples{
# 1-1-1 mediation with two random paths
model <- load_model('inst/examples/random_random')
simple_version(model, 's1', 's2', output='random_random')

}
