% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_model}
\alias{load_model}
\title{Title: Load Model Results from Mplus output file}
\usage{
load_model(file_name)
}
\arguments{
\item{filename}{filename or path of the Mplus input or output file}
}
\value{
Model results (list object), including fit statistics and parameters.
}
\description{
This function loads model reulst from Mplus output file.
If these exists no output file, this function will try to run the input file
 and reads the output file then.
The function depends on MplusAutomation package.
 https://www.statmodel.com/usingmplusviar.shtml
}
\examples{
model <- load_model('inst/examples/example_single')

}
