% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_mod.R
\name{med_mod}
\alias{med_mod}
\title{Title: Simple version of mediation test for mediated moderation test}
\usage{
med_mod(
  Model,
  iv = NULL,
  m = NULL,
  w = NULL,
  xw = NULL,
  dv = NULL,
  first_path = NULL,
  level_one_mod = TRUE,
  output = "Untitled",
  rep = 20000,
  conf = 95
)
}
\arguments{
\item{Model}{a list object containing the parameters of the model
(from Mplus output file)}

\item{iv}{the name of the independent variable}

\item{m}{the name of the mediator}

\item{w}{the name of the moderator}

\item{xw}{the name of the interaction term}

\item{dv}{the name of the dependent variable}

\item{first_path}{the name of the first random path (should not be 'fixed')}

\item{level_one_mod}{a binary variable indicating whether the moderation
effect is on the first level (default: TRUE)}

\item{output}{the name of the output image file (.png)}

\item{rep}{the number of repetitions (default: 20000)}

\item{conf}{the critical value used to generate confidence interval}
}
\value{
An image containing the distribution of the indirect effect and
confidence interval
}
\description{
The template of the monte carlo method in this function is from
 https://quantpsy.org/medmc/medmc.htm.
This function can handle multi-level mediated moderation model where the
 moderator can be specified at either level one or two.
}
