% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_modmed.R
\name{simple_modmed}
\alias{simple_modmed}
\title{Title: Moderated Mediation test of multilevel model (at least one path is fixed)
This function can calculate moderated mediation effect of multi-level model that
 contains at least one fixed path.}
\usage{
simple_modmed(
  Model,
  iv = NULL,
  m = NULL,
  w = NULL,
  xw = NULL,
  dv = NULL,
  first_stage_mod = TRUE,
  level_one_mod = TRUE,
  first_path = "fixed",
  second_path = "fixed",
  output = "Untitled",
  rep = 20000,
  conf = 95
)
}
\arguments{
\item{Model}{a list object containing the parameters of the model
(from Mplus output file)}

\item{iv}{the name of the independent variable}

\item{m}{the name of the mediator}

\item{w}{the name of the moderator}

\item{xw}{the name of the interaction term}

\item{dv}{the name of the dependent variable}

\item{first_stage_mod}{a binary variable indicating whether the moderation
effect is on the first path (default: TRUE)}

\item{level_one_mod}{a binary variable indicating whether the moderation
effect is on the first level (default: TRUE)}

\item{first_path}{the name of the first random path, left as 'fixed'
when the path is not random (default: 'fixed')}

\item{second_path}{the name of the second random path, left as 'fixed'
when the path is not random (default: 'fixed')}

\item{output}{the name of the output image file (.png)}

\item{rep}{the number of repetitions (default: 20000)}

\item{conf}{the critical value used to generate confidence interval}
}
\value{
An image containing the distribution of the indirect effect and
confidence interval
}
\description{
Title: Moderated Mediation test of multilevel model (at least one path is fixed)
This function can calculate moderated mediation effect of multi-level model that
 contains at least one fixed path.
}
