% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_version.R
\name{simple_version}
\alias{simple_version}
\title{Title: Simple version of mediation test}
\usage{
simple_version(
  Model,
  iv = NULL,
  m = NULL,
  dv = NULL,
  multi_level = FALSE,
  first_path = "fixed",
  second_path = "fixed",
  output = "Untitled",
  rep = 20000,
  conf = 95
)
}
\arguments{
\item{Model}{a list object containing the parameters of the model
(from Mplus output file)}

\item{iv}{the name of the independent variable}

\item{m}{the name of the mediator}

\item{dv}{the name of the dependent variable}

\item{multi_level}{a binary variable indicating whether the specified is
multi-level (default: FALSE)}

\item{first_path}{the name of the first random path, left as 'fixed'
when the path is not random (default: 'fixed')}

\item{second_path}{the name of the second random path, left as 'fixed'
when the path is not random (default: 'fixed')}

\item{output}{the name of the output image file (.png)}

\item{rep}{the number of repetitions (default: 20000)}

\item{conf}{the critical value used to generate confidence interval
(default: 95)}
}
\value{
An image containing the distribution of the indirect effect and
confidence interval
}
\description{
The template of the monte carlo method in this function is from
 https://quantpsy.org/medmc/medmc.htm.
This function can handle single level model and multi-level model that
 contains at least one fixed path.
}
\examples{
# Single level
model <- load_model('inst/examples/example_single')
simple_version(model, 'age', 'income', 'pain')

# 1-1-1 mediation with two fixed paths
model <- load_model('inst/examples/fixed_fixed')
simple_version(model, 'x', 'm', 'y')


}
